requirements.txt
akad==0.11.1
altgraph==0.16.1
astroid==2.2.5
attrs==19.1.0
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
Django==2.2.3
et-xmlfile==1.0.1
firebase==3.0.1
Flask==1.0.2
future==0.17.1
gunicorn==19.9.0
idna==2.8
imutils==0.5.2
inflect==2.1.0
ipython-genutils==0.2.0
isort==4.3.17
itsdangerous==1.1.0
jaraco.itertools==4.4.2
jdcal==1.4
Jinja2==2.10
jsonschema==3.0.1
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
line-bot-sdk==1.14.0
line-notify==0.1.4
linepy==3.0.8
lxml==4.3.3
macholib==1.11
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
more-itertools==7.2.0
nbformat==4.4.0
numpy==1.16.2
oauthlib==3.0.1
opencv-python==4.1.0.25
openpyxl==2.6.1
pafy==0.5.4
paho-mqtt==1.4.0
pandas==0.24.2
pefile==2019.4.18
Pillow==5.4.1
playsound==1.2.2
plotly==3.8.1
pyasn1==0.4.5
PyAutoGUI==0.9.42
pygame==1.9.6
PyGetWindow==0.0.4
PyInstaller==3.4
pylint==2.3.1
PyMsgBox==1.0.6
pyparsing==2.4.0
pypi==2.1
PyQRCode==1.2.1
PyRect==0.1.4
pyrsistent==0.14.11
PyScreeze==0.1.20
python-barcode==0.9.0
python-dateutil==2.8.0
python-firebase==1.2
python-firebase-gae==1.0.1
python-pptx==0.6.17
PyTweening==1.0.3
pytz==2019.1
pywin32==224
pywin32-ctypes==0.2.0
requests==2.3.0
requests-oauthlib==1.2.0
retrying==1.3.3
rsa==4.0
scipy==1.2.1
seaborn==0.9.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.1
spicy==0.16.0
sqlparse==0.3.0
thrift==0.11.0
traitlets==4.3.2
tweetpy==0.1
typed-ast==1.3.1
urllib3==1.24.1
vincent==0.4.4
virtualenv==16.6.2
WebOb==1.6.1
Werkzeug==0.15.1
wikipedia==1.4.0
wolframalpha==3.0.1
wrapt==1.11.1
XlsxWriter==1.1.7
xmltodict==0.12.0
youtube-dl==2019.3.18
